changed retry keyword to avoid conflict with phabricator
Summary:
This is a workaround, as our original pattern, !retry, conflicts with Phabricator's built-in mail commands.

In the future, if we want to invest time to go back to !retry, we should take a look at this file in the Phabricator repo.
src/applications/maniphest/command/ManiphestAssignEmailCommand.php
Phabricator mail commands works by finding all subclasses of MetaMTAEmailTransactionCommand

However, be really careful, as this is not a documented feature.

Test Plan: manual testing

Reviewers: josiah

Reviewed By: josiah

Differential Revision: https://tails.corp.dropbox.com/D119585