Allow changes to build dependent diffs
Summary:
When getting the base commit and patch on top of it, try to find parent diffs
and use those to build a patch from a landed commit to the current diff.

Test Plan:
Get a local version of Phabricator up and running, print out the generated
patch and see that it works for normal non-dependend diffs and basic dependent
diffs.

Test that it gracefully gives up when it encounters a depdendent diff where the
parent hasn't been diffed (Phabricator doesn't know about it) or a dependent
diff that's taken from the middle of a parent.

Reviewers: josiah

Reviewed By: josiah

Differential Revision: https://tails.corp.dropbox.com/D168607