Pass along list of diffs to changes.
Summary:
The main change here is an additional field, `phabricator.revisionIDList`,
which is passed along to Changes and contains an array of diff ids,
representing the diffs between master and this diff, when this is a dependent
diff. If this is not a dependent diff, this array will have a single item in
it.

Test Plan: doitlive

Reviewers: kylec, josiah

Reviewed By: josiah

Differential Revision: https://tails.corp.dropbox.com/D172769