Use the sourceControlBaseRevision instead of the oldest local commit when moving up diffs
Summary:
Currently, when generating patches for Changes, we traverse diffs by commit hash until we reach `master`. We assume that the Phabricator "raw diff" is between the parent of the oldest local commit and the newest local commit (by time). This assumption fails in a couple places:
 - First, if there are merges in the commit history, the base is often not the oldest commit.
 - If commits are not ordered by time (perhaps someone rebased something), the newest local commit may not be correct either.

Instead, we directly use Phabricator's `sourceControlBaseRevision` as the base, which is definitely the correct base, and we use the first local commit returned by Phabricator, which corresponds to HEAD.

Reviewers: josiah, alexmv

Reviewed By: alexmv

Differential Revision: https://tails.corp.dropbox.com/D222754