Make conduit calls as diff author
Summary:
For a clusterized repo, the conduit call performs an RPC to a machine
with the desired repository. Doing this as the omnipotent user means that
Phabricator tries to authenticate the request using the device's signing key,
making the assumption that it is executing in a daemon context (see
https://secure.phabricator.com/source/phabricator/browse/master/src/applications/repository/storage/PhabricatorRepository.php;5ff02058a4b0bcf1aa98a162a415ae7990ff14f5$2032).

In this case, however, our herald action is being executed by the web server
while handling an API request (creating or updating a revision). We do not have
the ability to read the device key (it is owned by the daemon user and mode
0600). To avoid trying to read the key, we'll perform the call as the user creating
the diff.

Test Plan:
Tested on a local phabricator VM. This didn't exercise the
clustered code path, but it confirms that we don't break the unclustered case.

Reviewers: alexmv

Reviewed By: alexmv

Differential Revision: https://tails.corp.dropbox.com/D256090